Re-organize CommitMessageBuilders
